requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
joblib==1.2.0
nltk==3.7
psycopg2==2.9.4
pytz==2022.4
regex==2022.9.13
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.4
whitenoise==6.2.0
gunicorn==20.0.4
